refactor(type-checker): extract Narrowing partial class [slice 3/6]
Move 9 narrowing methods from PreceptTypeChecker.cs into
PreceptTypeChecker.Narrowing.cs as part of issue #118 decomposition.

Methods moved:
- BuildEventEnsureSymbols
- BuildStateEnsureNarrowings
- BuildEventEnsureNarrowings
- ApplyAssignmentNarrowing
- ApplyNarrowing
- TryApplyNullComparisonNarrowing
- TryApplyNumericComparisonNarrowing
- TryStoreLinearFact
- TryDecomposeNullOrPattern

No behavior changes. Callers in main file and other partial files unchanged.
FlipComparisonOperator and TryGetNumericLiteral remain in main file (moved in later slices).

Closes #118 (partial)
